requirements.txt
flask
tensorflow>=2.10
pillow
imageio
numpy
